requirements.txt
fastapi[all]
firebase-admin==6.5.0
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart
uvicorn[standard]
